Changed the interface to just use the updateBookList function.
