+++ M02/Assignment/Assignment/app/src/main/java/com/example/entries/BookRepoInterface.kt
@@ -3,9 +3,5 @@ package com.example.entries
 import com.example.entries.model.Book
 
 interface BookRepoInterface {
-    fun saveAllBookCvs()
-    fun saveAllIds()
-    fun getAllBookIds(): String
-    fun getBookCSV(id: String): String
-    fun updateBook(book: Book)
+    fun updateBookList()
 }
